refactor(template-compiler): ignore invalid non-root template elements and log warning (#2977)
