--- packages/@lwc/integration-karma/test/template/empty-template/x/comment/comment.html
@@ -1,5 +1,0 @@
-<template lwc:preserve-comments>
-    <template class="foo">
-        <!--I am a comment-->
-    </template>
-</template>
