--- packages/@lwc/integration-karma/test/template/empty-template/x/comment/comment.js
@@ -1,3 +1,0 @@
-import { LightningElement } from 'lwc';
-
-export default class Comment extends LightningElement {}
