--- packages/@lwc/integration-karma/test/template/empty-template/x/mixed-directive/mixed-directive.html
@@ -1,7 +1,0 @@
-<template>
-    <template if:true={show} for:each={items} for:item="item">
-        <template class="foo">
-            <li key={item.key}>{item.value}</li>
-        </template>
-    </template>
-</template>
