--- packages/@lwc/integration-karma/test/template/empty-template/x/mixed/mixed.html
@@ -1,9 +1,0 @@
-<template>
-    <template class="foo">
-        before
-        <template class="foo">
-            <span>middle</span>
-        </template>
-        after
-    </template>
-</template>
