--- packages/@lwc/integration-karma/test/template/empty-template/x/mixed/mixed.js
@@ -1,3 +1,0 @@
-import { LightningElement } from 'lwc';
-
-export default class Mixed extends LightningElement {}
