--- packages/@lwc/integration-karma/test/template/empty-template/x/nested-element/nested-element.html
@@ -1,8 +1,0 @@
-<template>
-    <template class="foo">
-        <template class="foo">
-            <span>I am a span</span>
-        </template>
-    </template>
-</template>
-
