--- packages/@lwc/integration-karma/test/template/empty-template/x/nested-element/nested-element.js
@@ -1,3 +1,0 @@
-import { LightningElement } from 'lwc';
-
-export default class NestedElement extends LightningElement {}
