--- packages/@lwc/integration-karma/test/template/empty-template/x/nested-text/nested-text.html
@@ -1,8 +1,0 @@
-<template>
-    <template class="foo">
-        <template class="foo">
-            text
-        </template>
-    </template>
-</template>
-
