--- packages/@lwc/integration-karma/test/template/empty-template/x/nested-text/nested-text.js
@@ -1,3 +1,0 @@
-import { LightningElement } from 'lwc';
-
-export default class NestedText extends LightningElement {}
