--- packages/@lwc/integration-karma/test/template/empty-template/x/test/test.html
@@ -1,24 +1,0 @@
-<template>
-    <!-- Single text node -->
-    <template class="foo">Text</template>
-
-    <!-- Single element -->
-    <template class="foo">
-        <span>Span</span>
-    </template>
-
-    <!-- Multiple elements -->
-    <template class="foo">
-        <span>Span #1</span>
-        <span>Span #2</span>
-    </template>
-
-    <!-- Mixed -->
-    <template class="foo">
-        Before
-        <span>Span #1</span>
-        Middle
-        <span>Span #2</span>
-        After
-    </template>
-</template>
