--- packages/@lwc/integration-karma/test/template/empty-template/x/test/test.js
@@ -1,3 +1,0 @@
-import { LightningElement } from 'lwc';
-
-export default class Test extends LightningElement {}
