+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parsing-errors/invalid-template-attribute/enableStrictTemplateSyntax-flag-disabled/invalid-attribute-with-children/actual.html
@@ -1,0 +1,5 @@
+<template>
+    <template class="slds-m-around_medium">
+        Hello
+    </template>
+</template>
