+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parsing-errors/invalid-template-attribute/enableStrictTemplateSyntax-flag-disabled/invalid-attribute-with-children/metadata.json
@@ -1,0 +1,26 @@
+{
+    "warnings": [
+        {
+            "code": 1145,
+            "message": "LWC1145: Invalid attributes detected on template. The following attributes are not supported on template tags in LWC: class. For more information, please visit https://lwc.dev/guide/reference#html-template-directives",
+            "level": 2,
+            "location": {
+                "line": 2,
+                "column": 5,
+                "start": 15,
+                "length": 69
+            }
+        },
+        {
+            "code": 1153,
+            "message": "LWC1153: Non-root template elements must contain valid LWC template directive attributes. Otherwise, the template and its children will be ignored. For more information please visit https://lwc.dev/guide/reference#html-template-directives",
+            "level": 2,
+            "location": {
+                "line": 2,
+                "column": 5,
+                "start": 15,
+                "length": 69
+            }
+        }
+    ]
+}
