+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parsing-errors/invalid-template-attribute/enableStrictTemplateSyntax-flag-disabled/invalid-attribute/expected.js
@@ -1,7 +1,7 @@
 import { registerTemplate } from "lwc";
+const stc0 = [];
 function tmpl($api, $cmp, $slotset, $ctx) {
-  const { t: api_text } = $api;
-  return [api_text("Hello")];
+  return stc0;
   /*LWC compiler vX.X.X*/
 }
 export default registerTemplate(tmpl);
