+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parsing-errors/invalid-template-attribute/enableStrictTemplateSyntax-flag-disabled/invalid-attribute/metadata.json
@@ -10,6 +10,17 @@
                 "start": 15,
                 "length": 39
             }
+        },
+        {
+            "code": 1153,
+            "message": "LWC1153: Non-root template elements must contain valid LWC template directive attributes. Otherwise, the template and its children will be ignored. For more information please visit https://lwc.dev/guide/reference#html-template-directives",
+            "level": 2,
+            "location": {
+                "line": 2,
+                "column": 5,
+                "start": 15,
+                "length": 39
+            }
         }
     ]
 }
