+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parsing-errors/invalid-template-attribute/enableStrictTemplateSyntax-flag-disabled/mixed-valid-invalid-attributes/metadata.json
@@ -5,9 +5,9 @@
             "message": "LWC1145: Invalid attributes detected on template. The following attributes are not supported on template tags in LWC: onfoo. For more information, please visit https://lwc.dev/guide/reference#html-template-directives",
             "level": 2,
             "location": {
-                "line": 3,
+                "line": 4,
                 "column": 9,
-                "start": 33,
+                "start": 34,
                 "length": 142
             }
         }
