+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parsing-errors/invalid-template-attribute/enableStrictTemplateSyntax-flag-disabled/nested-invalid-attributes/actual.html
@@ -1,7 +1,9 @@
 <template>
     <section>
         <template if:true={showItems} for:each={items} for:item="item" onfoo={handleFoo}>
-            <p key={item.id}>1{item}</p>
+            <template class="slds-m-around_medium">
+                <p key={item.id}>1{item}</p>
+            </template>
         </template>
     </section>
 </template>
