+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parsing-errors/invalid-template-attribute/enableStrictTemplateSyntax-flag-disabled/nested-invalid-attributes/expected.js
@@ -1,0 +1,20 @@
+import { registerTemplate } from "lwc";
+const stc0 = {
+  key: 0,
+};
+const stc1 = [];
+function tmpl($api, $cmp, $slotset, $ctx) {
+  const { i: api_iterator, h: api_element } = $api;
+  return [
+    api_element(
+      "section",
+      stc0,
+      api_iterator($cmp.items, function (item) {
+        return stc1;
+      })
+    ),
+  ];
+  /*LWC compiler vX.X.X*/
+}
+export default registerTemplate(tmpl);
+tmpl.stylesheets = [];
