+++ packages/@lwc/template-compiler/src/__tests__/fixtures/parsing-errors/invalid-template-attribute/enableStrictTemplateSyntax-flag-disabled/nested-invalid-attributes/metadata.json
@@ -8,7 +8,7 @@
                 "line": 3,
                 "column": 9,
                 "start": 33,
-                "length": 575
+                "length": 222
             }
         },
         {
@@ -16,9 +16,20 @@
             "message": "LWC1145: Invalid attributes detected on template. The following attributes are not supported on template tags in LWC: class. For more information, please visit https://lwc.dev/guide/reference#html-template-directives",
             "level": 2,
             "location": {
-                "line": 11,
+                "line": 4,
                 "column": 13,
-                "start": 480,
+                "start": 127,
+                "length": 108
+            }
+        },
+        {
+            "code": 1153,
+            "message": "LWC1153: Non-root template elements must contain valid LWC template directive attributes. Otherwise, the template and its children will be ignored. For more information please visit https://lwc.dev/guide/reference#html-template-directives",
+            "level": 2,
+            "location": {
+                "line": 4,
+                "column": 13,
+                "start": 127,
                 "length": 108
             }
         }
