--- packages/@lwc/template-compiler/src/__tests__/fixtures/parsing-errors/invalid-template-attribute/multiple-attributes-mixed/actual.html
@@ -1,16 +1,0 @@
-<template>
-    <section>
-        <template if:true={showItems} for:each={items} for:item="item" onfoo={handleFoo}>
-            <!-- 
-                The current behavior is for the p element to be rendered to the DOM which is incorrect.
-                This will be replaced with an error message in the future but is kept here for now to preserve
-                backwards compatibility.
-                
-                Work to be done in W-11296404
-            -->
-            <template class="slds-m-around_medium">
-                <p key={item.id}>1{item}</p>
-            </template>
-        </template>
-    </section>
-</template>
