--- packages/@lwc/template-compiler/src/__tests__/fixtures/parsing-errors/invalid-template-attribute/multiple-attributes/expected.js
@@ -1,33 +1,0 @@
-import { registerTemplate } from "lwc";
-const stc0 = {
-  key: 0,
-};
-function tmpl($api, $cmp, $slotset, $ctx) {
-  const {
-    k: api_key,
-    d: api_dynamic_text,
-    t: api_text,
-    h: api_element,
-    i: api_iterator,
-  } = $api;
-  return [
-    api_element(
-      "section",
-      stc0,
-      api_iterator($cmp.items, function (item) {
-        return $cmp.showItems
-          ? api_element(
-              "p",
-              {
-                key: api_key(1, item.id),
-              },
-              [api_text("1" + api_dynamic_text(item))]
-            )
-          : null;
-      })
-    ),
-  ];
-  /*LWC compiler vX.X.X*/
-}
-export default registerTemplate(tmpl);
-tmpl.stylesheets = [];
