--- packages/@lwc/template-compiler/src/__tests__/fixtures/parsing-errors/invalid-template-attribute/single-attribute-with-children/actual.html
@@ -1,12 +1,0 @@
-<template>
-    <!-- 
-        The current behavior is for the text to be rendered to the DOM which is incorrect.
-        This will be replaced with an error message in the future but is kept here for now to preserve
-        backwards compatibility.
-
-        Work to be done in W-11296404
-    -->
-    <template class="slds-m-around_medium">
-        Hello
-    </template>
-</template>
