--- packages/@lwc/template-compiler/src/__tests__/fixtures/parsing-errors/invalid-template-attribute/single-attribute-with-children/metadata.json
@@ -1,15 +1,0 @@
-{
-    "warnings": [
-        {
-            "code": 1145,
-            "message": "LWC1145: Invalid attributes detected on template. The following attributes are not supported on template tags in LWC: class. For more information, please visit https://lwc.dev/guide/reference#html-template-directives",
-            "level": 2,
-            "location": {
-                "line": 9,
-                "column": 5,
-                "start": 299,
-                "length": 69
-            }
-        }
-    ]
-}
